Test : moved test to an other file, deleted one useless function
