+++ app/src/test/java/com/android/feedme/ui/TextRecognitionTest.kt
@@ -13,6 +13,7 @@ import com.google.mlkit.vision.text.Text
 import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.Request
 import okhttp3.RequestBody.Companion.toRequestBody
+import org.json.JSONObject
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.mockito.Mockito.mock
@@ -124,5 +125,46 @@ class TextRecognitionTest {
 
     // Assert that the result contains the text from the block
     assert(result == "This is a test block")
+  }
+
+  @Test
+  fun testParseResponse() {
+    // Create a mock JSON response
+    val jsonResponse =
+      JSONObject(
+        """
+            {
+                "choices": [
+                    {
+                        "message": {
+                            "content": "[{\"ingredient\":\"flour\",\"quantity\":\"1\",\"unit\":\"cup\"},{\"ingredient\":\"eggs\",\"quantity\":\"2\",\"unit\":\"\"},{\"ingredient\":\"salt\",\"quantity\":\"0.5\",\"unit\":\"teaspoon\"}]"
+                        }
+                    }
+                ]
+            }
+        """
+          .trimIndent())
+
+    // Capture the callback arguments
+    val ingredientsList = mutableListOf<IngredientMetaData>()
+    val forIngredientFound: (IngredientMetaData) -> Unit = { ingredientsList.add(it) }
+
+    // Call the function under test
+    parseResponse(jsonResponse.toString(), forIngredientFound)
+
+    // Verify the captured arguments
+    assert(ingredientsList.size == 3)
+    assert(
+      ingredientsList[0].quantity == 1.0 &&
+              ingredientsList[0].measure == MeasureUnit.CUP &&
+              ingredientsList[0].ingredient.name == "Flour")
+    assert(
+      ingredientsList[1].quantity == 2.0 &&
+              ingredientsList[1].measure == MeasureUnit.NONE &&
+              ingredientsList[1].ingredient.name == "Eggs")
+    assert(
+      ingredientsList[2].quantity == 0.5 &&
+              ingredientsList[2].measure == MeasureUnit.TEASPOON &&
+              ingredientsList[2].ingredient.name == "Salt")
   }
 }
